Fixes some GetXXXXs functions
For iterations over a range will reuse the same variable, so adding its
address to a slice will effectively add the same element over and over
again.
Making a copy before adding the element will ensure that all the
elements are passed.
Added a few tests to verify all the GetXXXXs functions