fix bug in `GetSchedules`
While making my lights more awesome, I noticed a bug in `GetSchedules` that caused all items in the `return` array to be the same `*Schedule`.

From this, I was able to derive [this minimal repo](https://play.golang.org/p/u6cbj8jQNDr).

As best as I can tell, the `append` is coping the pointer to the `Schedule` which is being reset as per the `range`.

There may be a more elegant fix, but simply assigning `s` to another variable and referencing that [works here](https://play.golang.org/p/hgTFJ2Mcwa4)

While I was here, I noticed that the results were not ordered as go maps don't provide that. Since the Hue api returns ordered results, I made the behavior constient.