Update schedule.go (#14)
Currently, I am unable to create a recurring schedule. It results in the following error: 
```
panic: ERROR 6 [/schedules]: "parameter, autodelete, not available"
```
Based on what I am seeing for the `autodelete` param in the (schedule docs)[https://developers.meethue.com/develop/hue-api/3-schedules-api/#create-schedule], this should not be passed if the schedule is reccouring. 

I used `curl` to confirm this hypothesis. 

I am now testing if setting `omitempty` on the `AutoDelete` field of the `Schedule` struct will solve this issue.

Co-authored-by: Amir Mofasser <amir.mofasser@gmail.com>